\name{print.3vel}
\alias{print.3vel}
\title{Print methods for three-velocities}
\description{Print methods for three-velocities}
\usage{
\method{print}{3vel}(x, ...)
}
\arguments{
  \item{x}{Vector of three-velocities}
  \item{\dots}{Further arguments, currently ignored}
}
\value{Returns a vector of three-velocities}
\author{Robin K. S. Hankin}
\examples{
r3vel(10)
}